+++ test/dtypes/test_affine_quantized_float.py
@@ -26,11 +26,15 @@ from torchao.quantization import (
     float8_weight_only,
     quantize_,
 )
-from torchao.quantization.observer import PerRow, PerTensor
 from torchao.quantization.quant_api import (
     float8_static_activation_float8_weight,
 )
-from torchao.quantization.quant_primitives import MappingType, choose_qparams_affine
+from torchao.quantization.quant_primitives import (
+    choose_qparams_affine,
+    MappingType,
+    PerRow,
+    PerTensor,
+)
 
 random.seed(0)
 torch.manual_seed(0)
