+++ tutorials/calibration_flow/awq_like.py
@@ -22,11 +22,11 @@ from torchao.quantization import to_weight_tensor_with_linear_activation_scale_m
 from torchao.quantization.quant_api import _replace_with_custom_fn_if_matches_filter
 from torchao.quantization.observer import (
     AffineQuantizedMinMaxObserver,
-    PerTensor,
-    PerAxis,
 )
 from torchao.quantization.quant_primitives import (
     MappingType,
+    PerTensor,
+    PerAxis,
     FP8_TYPES,
 )
