+++ tutorials/calibration_flow/static_quant.py
@@ -17,11 +17,11 @@ from torchao.quantization import to_linear_activation_quantized
 from torchao.quantization.quant_api import _replace_with_custom_fn_if_matches_filter
 from torchao.quantization.observer import (
     AffineQuantizedMinMaxObserver,
-    PerTensor,
-    PerAxis,
 )
 from torchao.quantization.quant_primitives import (
     MappingType,
+    PerTensor,
+    PerAxis,
     FP8_TYPES, 
 )
